+++ packages/netlify-cms-app/src/index.js
@@ -3,6 +3,15 @@ import './backends';
 import './widgets';
 import './editor-components';
 
+if (typeof window !== 'undefined') {
+  /**
+   * Log the version number.
+   */
+  if (typeof NETLIFY_CMS_APP_VERSION === 'string') {
+    console.log(`netlify-cms-app ${NETLIFY_CMS_APP_VERSION}`);
+  }
+}
+
 export const NetlifyCmsApp = {
   ...CMS,
 };