+++ packages/netlify-cms-app/webpack.config.js
@@ -1,4 +1,3 @@
-const path = require('path');
 const webpack = require('webpack');
 const pkg = require('./package.json');
 const { getConfig, plugins } = require('../../scripts/webpack');
@@ -9,15 +8,12 @@ console.log(`${pkg.version}${isProduction ? '' : '-dev'}`);
 
 const baseConfig = {
   ...baseWebpackConfig,
-  context: path.join(__dirname, 'src'),
-  entry: './index.js',
   plugins: [
     ...Object.entries(plugins)
       .filter(([key]) => key !== 'friendlyErrors')
       .map(([, plugin]) => plugin()),
     new webpack.DefinePlugin({
-      NETLIFY_CMS_VERSION: JSON.stringify(`- app - ${pkg.version}${isProduction ? '' : '-dev'}`),
-      NETLIFY_CMS_CORE_VERSION: null,
+      NETLIFY_CMS_APP_VERSION: JSON.stringify(`${pkg.version}${isProduction ? '' : '-dev'}`),
     }),
   ],
 };