+++ packages/netlify-cms-core/webpack.config.js
@@ -5,7 +5,6 @@ const { getConfig } = require('../../scripts/webpack.js');
 const isProduction = process.env.NODE_ENV === 'production';
 
 const versionPlugin = new webpack.DefinePlugin({
-  NETLIFY_CMS_VERSION: null,
   NETLIFY_CMS_CORE_VERSION: JSON.stringify(`${pkg.version}${isProduction ? '' : '-dev'}`),
 });